requirements.txt
aiofiles==24.1.0
aiohttp_socks==0.9.0
aiohttp==3.10.10
colorama==0.4.6
ddddocr==1.5.6
fake_useragent==1.5.1
loguru==0.7.2
parsel==1.9.1
Pillow==11.0.0
python-dotenv==1.0.1
PyYAML==6.0.2
Requests==2.32.3
termcolor==2.5.0
urllib3==2.2.3